button to see password (see if flutter don't have one already) Fixes #26
