+++ mobile/account_managment/lib/screens/login_screen.dart
@@ -1,58 +1,16 @@
-import 'package:account_managment/viewModels/auth_view_model.dart';
+import 'package:account_managment/forms/login_form.dart';
 import 'package:flutter/material.dart';
-import 'package:provider/provider.dart';
 
 class LoginScreen extends StatelessWidget {
-  final TextEditingController usernameController = TextEditingController();
-  final TextEditingController passwordController = TextEditingController();
-
-  LoginScreen({super.key});
+  const LoginScreen({super.key});
 
   @override
   Widget build(BuildContext context) {
-    final authViewModel = Provider.of<AuthViewModel>(context);
-
     return Scaffold(
       appBar: AppBar(title: const Text("Login")),
-      body: Padding(
-        padding: const EdgeInsets.all(16.0),
-        child: Column(
-          children: [
-            TextField(
-              controller: usernameController,
-              decoration: const InputDecoration(labelText: 'Username'),
-            ),
-            TextField(
-              controller: passwordController,
-              decoration: const InputDecoration(labelText: 'Password'),
-              obscureText: true,
-            ),
-            const SizedBox(height: 16),
-            ElevatedButton(
-              onPressed: () async {
-                await authViewModel.login(
-                  usernameController.text,
-                  passwordController.text,
-                );
-                if (authViewModel.accessToken != null) {
-                  Navigator.pushReplacementNamed(context, '/accounts');
-                } else {
-                  ScaffoldMessenger.of(context).showSnackBar(
-                    const SnackBar(content: Text("Invalid credentials")),
-                  );
-                }
-              },
-              child: const Text("Login"),
-            ),
-            const SizedBox(height: 16),
-            ElevatedButton(
-              onPressed: () {
-                Navigator.pushNamed(context, '/profile');
-              },
-              child: const Text("Create account"),
-            ),
-          ],
-        ),
+      body: const Padding(
+        padding: EdgeInsets.all(16.0),
+        child: LoginForm(),
       ),
     );
   }
